requirements.txt
pandas
numpy
Pillow
streamlit
matplotlib
plotly
openpyxl
xlrd
pyxlsb
